Move utils from optimisation to right under qamomile
