+++ qamomile/optimization/binary_model/expr.py
@@ -5,7 +5,7 @@ import enum
 from collections import Counter
 from typing import Generic, TypeVar
 
-from qamomile.optimization.utils import is_close_zero
+from qamomile._utils import is_close_zero
 
 
 class VarType(enum.StrEnum):
