+++ qamomile/optimization/fqaoa.py
@@ -39,6 +39,7 @@ import ommx.v1
 
 import qamomile.circuit as qmc
 import qamomile.observable as qm_o
+from qamomile._utils import is_close_zero
 from qamomile.circuit.algorithm.fqaoa import (
     fqaoa_state,
 )
@@ -46,7 +47,6 @@ from qamomile.circuit.transpiler.executable import ExecutableProgram
 from qamomile.circuit.transpiler.transpiler import Transpiler
 from qamomile.optimization.binary_model import BinaryModel
 from qamomile.optimization.converter import MathematicalProblemConverter
-from qamomile.optimization.utils import is_close_zero
 
 
 class FQAOAConverter(MathematicalProblemConverter):
