+++ qamomile/optimization/qrao/base_encoder.py
@@ -18,8 +18,8 @@ import abc
 from typing import Literal
 
 import qamomile.observable as qm_o
+from qamomile._utils import is_close_zero
 from qamomile.optimization.binary_model import BinaryModel, VarType
-from qamomile.optimization.utils import is_close_zero
 
 from .graph_coloring import check_linear_term, greedy_graph_coloring
